+++ CHANGELOG.md
@@ -40,7 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [x] TestAdapter - Executor
 [x] TestAdapter - Logger
 [x] TestAdapter - TestCaseUtils
-[x] TestAdapter - RunSettings (AdapterSettings, XmlExtension, RunSettingModel)
+[x] TestAdapter - RunSettings (AdapterSettings, XmlExtension, RunSettingModel, XmlBool)
 [x] Services - RevitTestConsole
 [x] Services - ProcessStart
 [x] Test Models